Implemented Simplified Solver
Added the simplified solver to the pos_solver file. It builds a dictionary of occurrences for each word - e.g. "swing" appeared as a noun 10 times and as a verb 15 times. It then picks the occurrence with the highest count as the part of speech to return (this is the same as taking argmax of occurrence for POS / total occurrences of word).

Accuracy: 93.85% for words, 47.50% for sentences on the full test dataset.