Early attempt at HMM_VE
Implemented an early attempt to using a HMM for the model. All of the pieces are there (initial states, emission probabilities, and transition probabilities). It runs, however, it's predictive power on both words and sentences are lower than the simplified version which just uses P(POS | Word). I'll do some debugging but wanted to get an initial commit in.