Renamed variables to HMM friendly ones.
Did a little renaming while I was thinking about viterbi. Open issue is how to deal with probabilities of zero - can use Laplace smoothing but need to not do that to the HMM-VE model is it degrades performance. Will likely just keep a second copy for the Viterbi during training in next update tomorrow.