Updated Emission Probabilities
Updated the HMM-VE algorithm to only attempt to apply the emission probabilities if the word exists in the Lexicon. If it hasn't, then the probabilities are just derived from the transition probabilities. Ends up with 93.41%/44.70% correct words and sentences. Still think I'm missing something that should be an improvement over the simplified version.